Fix for profile image corner radius on photo details view
In the photo details view, the user's profile image was blank because the corner radius of the PAPProfileImageView control was set to a fixed 66px, while the profile image size was only 35px. The expected effect was to have a round profile picture, so the code has been changed to fix this.